+++ src/main/java/org/webbitserver/stub/StubHttpRequest.java
@@ -2,10 +2,12 @@ package org.webbitserver.stub;
 
 import org.webbitserver.HttpRequest;
 import org.webbitserver.InboundCookieParser;
+import org.webbitserver.helpers.QueryParameters;
 
 import java.net.HttpCookie;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
+import java.net.URI;
 import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.List;
@@ -75,6 +77,16 @@ public class StubHttpRequest extends StubDataHolder implements HttpRequest {
             }
         }
         return null;
+    }
+
+    @Override
+    public String queryParam(String key) {
+        return new QueryParameters(URI.create(uri()).getQuery()).first(key);
+    }
+
+    @Override
+    public List<String> queryParams(String key) {
+        return new QueryParameters(URI.create(uri()).getQuery()).all(key);
     }
 
     @Override