+++ HISTORY.md
@@ -3,6 +3,7 @@
 
 * Requests with full url as request uri (and not only path) are correctly matched. (Aslak Hellesøy)
 * Added new AliasHandler for forwarding requests. (Aslak Hellesøy)
+* Added HttpRequest.queryParam(String key) and HttpRequest.queryParams(String key) (#20 Aslak Hellesøy)
 * Made it possible to call NettyHttpResponse.content() more than once. (Aslak Hellesøy)
 * Added support for cookies. (#19 Aslak Hellesøy)
 * Added HttpRequest.body() method. (Aslak Hellesøy)